feat: merge mobile/desktop AgentHome, auto-collapse chat at mobile width
- Remove AgentHomeDesktop/AgentHomeMobile split — single component
- Auto-collapse chat when hitting mobile breakpoint, uncollapse on resize
- Dashboard goes full-width when chat collapsed (no px-page on wrapper)
- Pass fullscreen param to dashboard iframe for responsive padding
- Fix useIsMobile initial state to avoid hydration flash
