+++ app/src/pages/AgentDashboard/index.tsx
@@ -1,15 +1,16 @@
+import { useOutletContext } from "react-router-dom";
 import { AgentHome } from "@/components/AgentHome";
 import { useLayout } from "@/stores/use-layout";
-import { useIsMobile } from "@/hooks/use-mobile";
+import type { AgentHomeOutletContext } from "@/lib/types";
 
 export function AgentDashboard() {
   const navbarHeight = useLayout((s) => s.navbarHeight);
-  const isMobile = useIsMobile();
+  const { chatCollapsed } = useOutletContext<AgentHomeOutletContext>();
 
   return (
     <div
-      className={`flex flex-col flex-1 min-h-0 relative overflow-hidden ${isMobile ? "" : "px-page"}`}
-      style={{ paddingTop: navbarHeight }}
+      className={`flex flex-col flex-1 min-h-0 relative overflow-hidden ${chatCollapsed ? "" : "px-page"}`}
+      style={{ paddingTop: `calc(${navbarHeight}px + var(--page-padding-x) / 1.5)` }}
     >
       <AgentHome />
     </div>
